pyproject.toml
[tool.poetry]
name = "language_assistant"
version = "0.1.0"
description = ""
authors = ["dmytro <dimatsaryuk30@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13.5"
openai = ">=1.90.0,<2.0.0"
streamlit = ">=1.46.0,<2.0.0"
sqlmodel = ">=0.0.24,<0.0.25"
sqlite-utils = ">=3.38,<4.0"
python-dotenv = ">=1.1.0,<2.0.0"
backoff = ">=2.2.1,<3.0.0"
ollama = ">=0.2.0,<1.0.0"
faiss-cpu = ">=1.7.0,<2.0.0"
sentence-transformers = ">=2.0.0,<3.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
